QuickSettings: Fix Swedish translations
The translation of "Bluetooth Off" to "Inaktivera Bluetooth"
(which means "Disable Bluetooth") causes confusion as the string
is meant to show the status, not what would happen if you tapped it.

The translation for "WiFi Off" is valid but it doesn't fit. Changed
to "WiFi Av" to match the change above. This one can be discussed.

The translation of "WiFi Display" translates the word "display"
as a verb which made no sense in this context. There is currently
no proper Swedish name for the WiFi Display standard so the English
name is used instead. It's what people actually use anyway. "WiFi visas"
means "WiFi is displayed" which is clearly wrong in this context.

Change-Id: I6013421242ee1dea6bbc2bc6adbf83581b4c6de2