Base: Allow hiding notification icons
When long pressing a notification, add another menu
item which allows the user to hide the notification icon
(but keep the actual notification in the dropdown).

This is meant for applications which have services running
in the statusbar which the user is aware of, so right now
we hide the icon by the package.

Video preview done by Hernán Castañón Álvarez running on MoKee:

http://youtu.be/jvTJmppGmJ4

Change-Id: If7499ee8d5ba7da00782b2573283f8c389ff6e51