SystemUI: Use Stock Android Colors instead of Dark Pixel Colors
I think it's better, when we use the Stock Android Primary Colors
instead of the Pixel Primary because right now,
we have Primary Pixel and Stock Accent Colors.

These Stock Colors were been used in mkl and also in mkm,
why not in mkn-mr1. It looks more clean with these Colors.
Many Rom’s switched to the Pixel look,
LineageOS is one of the last Custom Rom’s with Android 7.1.1
and an Half-Stock look. The Stock look would be Perfect for our Rom.
It’s not a Big Change but a Good Change.

[LineageOS right now: http://i.imgur.com/8Cujj0q.png]
[LineageOS after this Color Change: http://i.imgur.com/M7wA20X.png]

Change-Id: I0901811a4ba9a7d1d459689bfd333271e31204cb